+++ build.gradle.kts
@@ -26,7 +26,7 @@ val neoforge: SourceSet by sourceSets.creating {
 dependencies {
     implementation("org.jetbrains:annotations:24.1.0")
     implementation("com.google.code.gson:gson:2.10")
-    implementation("org.java-websocket:Java-WebSocket:1.6.0")
+    testImplementation("org.java-websocket:Java-WebSocket:1.6.0")
 
     "fabricImplementation"("net.fabricmc:fabric-loader:0.15.0") { isTransitive = false}
