+++ src/main/java/tech/thatgravyboat/repolib/v2/builtin/BuiltinArray.java
@@ -1,5 +1,6 @@
 package tech.thatgravyboat.repolib.v2.builtin;
 
+import tech.thatgravyboat.repolib.v2.expl.value.ArrayValue;
 import tech.thatgravyboat.repolib.v2.expl.value.KeyValue;
 import tech.thatgravyboat.repolib.v2.expl.value.MutableArrayValue;
 
@@ -10,19 +11,27 @@ import java.util.stream.StreamSupport;
 public class BuiltinArray {
 
 
-
     public static final KeyValue ARRAY = new Constants(builder -> {
 
-        builder.function("sorted", function -> {
-            function.arity(1);
-            function.execute((evaluator, args) -> {
-                var array = evaluator.getArrayOrThrow(args.getFirst());
-                var sorted = MutableArrayValue.create();
-                StreamSupport.stream(array.spliterator(), false).sorted(Comparator.comparing(Function.identity())).forEach(sorted::add);
+        builder.function(
+            "sorted", function -> {
+                function.arity(1);
+                function.execute((evaluator, args) -> {
+                    var array = evaluator.getArrayOrThrow(args.getFirst());
+                    var sorted = MutableArrayValue.create();
+                    StreamSupport.stream(array.spliterator(), false)
+                        .sorted(Comparator.comparing(Function.identity()))
+                        .forEach(sorted::add);
 
-                return sorted;
+                    return sorted;
+                });
             });
-        });
+
+        builder.function(
+            "of", function -> {
+                function.vararg(true);
+                function.execute((evaluator, args) -> MutableArrayValue.create(ArrayValue.flatten(args)));
+            });
 
     });
