+++ src/test/java/Main.java
@@ -1,5 +1,5 @@
-package tech.thatgravyboat.repolib.v2;
-
+import tech.thatgravyboat.repolib.v2.RepoConstants;
+import tech.thatgravyboat.repolib.v2.RepoLoader;
 import tech.thatgravyboat.repolib.v2.builtin.Constants;
 import tech.thatgravyboat.repolib.v2.expl.value.StructValue;
