+++ src/test/java/utils/WatchDir.java
@@ -1,4 +1,4 @@
-package tech.thatgravyboat.repolib.v2.utils;
+package utils;
 
 import java.io.IOException;
 import java.nio.file.FileSystems;
@@ -6,7 +6,6 @@ import java.nio.file.FileVisitResult;
 import java.nio.file.Files;
 import java.nio.file.LinkOption;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.nio.file.SimpleFileVisitor;
 import java.nio.file.StandardWatchEventKinds;
 import java.nio.file.WatchEvent;
